Extract Proxy logic to a different package, so it's reusable.
This allow us to use this logic in other places, rather than being
coupled to the http server.
