+++ .gitignore
@@ -1,3 +1,4 @@
 *.iml
 config.json
 vendor/
+gotiator
